MAINT: clean single purpose surface generation out of lib
