+++ a_package/model/__init__.py
@@ -1,4 +1,4 @@
 from .capillary import NodalFormCapillary
 from .contact import RigidContact
-from .surfaces import generate_surface, SelfAffineRoughness, psd_to_height
+from .roughness import SelfAffineRoughness, psd_to_height
 from .term import Term
