+++ cases/surface_helpers.py
@@ -1,18 +1,44 @@
 """
-Surface geometry generators.
+Surface geometry generators for simple shapes.
 
 Each generator creates a height field on a given grid based on surface parameters.
-This module is independent of config - it only works with primitives (shape string + params).
 """
 
-import dataclasses as dc
+from typing import Any
 
 import numpy as np
-import numpy.linalg as la
-import numpy.fft as fft
-import numpy.random as random
 
-from a_package.domain import Grid, Field, field_component_ax
+from a_package.domain import Grid
+
+
+def get_surface_shape(config: dict, which: str) -> str:
+    """
+    Get the shape name for a surface.
+
+    Parameters
+    ----------
+    config : dict
+        The configuration dict.
+    which : str
+        Either "upper" or "lower".
+
+    Returns
+    -------
+    str
+        The surface shape name.
+    """
+    return config["problem"][which]["shape"]
+
+
+def generate_surface_from_config(grid: Grid, surface_cfg: dict[str, Any]) -> np.ndarray:
+    """
+    Generate a surface from configuration dict.
+
+    Extracts shape and passes remaining params to generate_surface.
+    """
+    cfg = dict(surface_cfg)  # copy to avoid mutation
+    shape = cfg.pop("shape")
+    return generate_surface(grid, shape, **cfg)
 
 
 # Registry of surface generators
@@ -27,53 +53,6 @@ def _register(shape: str):
     return decorator
 
 
-@dc.dataclass(init=True)
-class SelfAffineRoughness:
-    C0: float
-    """Prefactor"""
-    qR: float
-    """Roll-off (angular) wavenumber"""
-    qS: float
-    """Cut-off (angular) wavenumber"""
-    H: float
-    """Hurst exponent"""
-
-    def mapto_isotropic_psd(self, q: Field):
-        """
-        Get the isotropic power spectral density (psd) of a given wavenumber
-        - q: wavenumber in radius, i.e. 2 pi over wavelength
-        """
-        # isotropic, only the magnitude matters
-        wavenumber = la.norm(q, ord=2, axis=field_component_ax, keepdims=True)
-
-        # Find three regimes
-        constant = wavenumber < self.qR
-        self_affine = (wavenumber >= self.qR) & (wavenumber < self.qS)
-        omitted = wavenumber >= self.qS
-
-        # Evaluate accordingly
-        psd = np.full_like(wavenumber, np.nan)
-        psd[constant] = self.C0 * self.qR ** (-2 - 2 * self.H)
-        psd[self_affine] = self.C0 * wavenumber[self_affine] ** (-2 - 2 * self.H)
-        psd[omitted] = 0
-
-        # Return both in convenience of plotting
-        return wavenumber, psd
-
-
-def psd_to_height(psd: Field, rng=None, seed=None):
-    # <h^2> corresponding to <PSD>, thus, take the square-root to match overall amplitude
-    h_amp = np.sqrt(psd)
-
-    # impose some random phase angle following uniform distribution
-    if rng is None:
-        rng = random.default_rng(seed)
-    phase_angle = np.exp(1j * rng.uniform(0, 2 * np.pi, psd.shape))
-
-    # only the sinusoidal is needed
-    return fft.ifft2(h_amp * phase_angle).real
-
-
 def generate_surface(grid: Grid, shape: str, **params) -> np.ndarray:
     """
     Generate a surface height field based on shape and parameters.
@@ -83,7 +62,7 @@ def generate_surface(grid: Grid, shape: str, **params) -> np.ndarray:
     grid : Grid
         The computational grid.
     shape : str
-        Surface type identifier ("flat", "tip", "sinusoid", "rough", "pattern").
+        Surface type identifier ("flat", "tip", "sinusoid", "pattern").
     **params
         Shape-specific parameters.
 
@@ -129,35 +108,8 @@ def _generate_tip(grid: Grid, radius: float) -> np.ndarray:
 def _generate_sinusoid(grid: Grid, wavenumber: float, amplitude: float) -> np.ndarray:
     """Generate a sinusoidal surface."""
     [x, y] = grid.form_nodal_mesh()
-    [qx, qy] = grid.form_spectral_mesh()
-    height = amplitude * np.cos(qx * x) * np.cos(qy * y)
+    height = amplitude * np.cos(wavenumber * x) * np.cos(wavenumber * y)
     return height
-
-
-@_register("rough")
-def _generate_rough(
-    grid: Grid,
-    prefactor: float,
-    rolloff_wavelength_pixels: int,
-    cutoff_wavelength_pixels: int,
-    hurst_exponent: float,
-    seed: int | None = None,
-) -> np.ndarray:
-    """Generate a self-affine rough surface from PSD."""
-    qR = (2 * np.pi) / (grid.element_sizes[0] * rolloff_wavelength_pixels)
-    qS = (2 * np.pi) / (grid.element_sizes[0] * cutoff_wavelength_pixels)
-
-    roughness = SelfAffineRoughness(prefactor, qR, qS, hurst_exponent)
-    q_2D = grid.form_spectral_mesh()
-    [_, C_2D] = roughness.mapto_isotropic_psd(q_2D)
-
-    # Get or generate the seed
-    seq = random.SeedSequence(seed)
-
-    # Generate rough surface from PSD
-    rng = random.default_rng(seq)
-    height = psd_to_height(C_2D, rng=rng)
-    return height.squeeze(axis=0)
 
 
 @_register("pattern")
